requirements.txt
aiohttp>=3.10.0
PyYAML>=6.0.1
eth-account>=0.11.3
web3>=6.20.0
py-clob-client>=0.34.0
